+++ src/ur_decoder.h
@@ -129,14 +129,14 @@ size_t ur_decoder_received_parts_count(ur_decoder_t *decoder);
  * @param decoder Pointer to URDecoder instance
  * @return Completion percentage (0.0 to 1.0)
  */
-double ur_decoder_estimated_percent_complete(ur_decoder_t *decoder);
+float ur_decoder_estimated_percent_complete(ur_decoder_t *decoder);
 
 /**
  * Get estimated completion percentage using the weighted-mixed-frames method.
  * @param decoder Pointer to URDecoder instance
  * @return Completion percentage (0.0 to 1.0)
  */
-double ur_decoder_estimated_percent_complete_weighted(ur_decoder_t *decoder);
+float ur_decoder_estimated_percent_complete_weighted(ur_decoder_t *decoder);
 
 void ur_result_free(ur_result_t *result);
